trusted-types/trusted-types-navigation.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/navigation-support.js"></script>
</head>
<body>
<script>
  "use strict";

  function expectMessage(filter) {
    return new Promise(resolve => {

      function listener(e) {
        if (filter(e)) {
          window.removeEventListener("message", listener);
          resolve();
        }
      }
      window.addEventListener("message", listener);
    });
  }

  function expectViolationAsMessage(sample) {
    const filter = e => {
      const result = (e.data.effectiveDirective == "require-trusted-types-for" &&
                         (!sample || e.data.sample.startsWith(sample)));
      if (result) {
        assert_true(true, "Expected violation as message: " + sample);
      }

      return result;
    }
    return new expectMessage(filter);
  }

  function expectLoadedAsMessage(searchParams, originAndPathName) {
    const filter = e => {
      if (e.data.type != "DOMContentLoaded") {
        return false;
      }
      let url = new URL(e.data.uri);
      if (!url.pathname.endsWith(originAndPathName)) {
        return false;
      }
      let keys = Object.keys(searchParams);
      if (url.searchParams.size != keys.length ||
          !keys.every(key =>
            url.searchParams.get(key) === searchParams[key].toString())) {
        return false;
      }

      assert_true(true, `Expected loaded as message: ${joinToHref(searchParams, originAndPathName)}`);
      return true;
    }
    return new expectMessage(filter);
  }

  function openWindow(test, uri) {
    const win = window.open(uri);
    test.add_cleanup(_ => win.close());
  }

  const kFormSubmission = "form-submission";
  // When adding more elements, adapt all functions consuming the existing elements.
  const kNavigationElements =
  [
    "anchor",
    kFormSubmission,
  ];

  function maybeAddFormSubmissionToSearchParams(navigationElement, searchParams) {
    return Object.assign(
      navigationElement == kFormSubmission ? {kFormSubmission: 1} : {},
      searchParams);
  }

  function joinToHref(searchParams, originAndPathName) {
    let urlSearchParams = new URLSearchParams(searchParams);
    if (urlSearchParams.size > 0) {
      return originAndPathName + "?" + urlSearchParams.toString();
    }

    return originAndPathName;
  }

  const kNavigationSupport = "navigation-support.html";
  const kNavigationSupportReportOnly = "navigation-report-only-support.html";

  for (const navigationElement of kNavigationElements) {
    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {});
      openWindow(t, `support/${joinToHref(params, kNavigationSupport)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupport),
        expectViolationAsMessage("Location href"),
      ]);
    }, `Navigate a window via ${navigationElement} with javascript:-urls in enforcing mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {defaultpolicy: 1});
      openWindow(t, `support/${joinToHref(params, kNavigationSupport)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupport),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, defaultpolicywashere: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a window via ${navigationElement} with javascript:-urls w/ default policy in enforcing mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {});
      openWindow(t, `support/${joinToHref(params, kNavigationSupportReportOnly)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupportReportOnly),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, continue: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a window via ${navigationElement} with javascript:-urls in report-only mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {defaultpolicy: 1});
      openWindow(t, `support/${joinToHref(params, kNavigationSupportReportOnly)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupportReportOnly),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, defaultpolicywashere: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a window via ${navigationElement} with javascript:-urls w/ default policy in report-only mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {frame: 1});
      openWindow(t, `support/${joinToHref(params, kNavigationSupport)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupport),
        expectViolationAsMessage("Location href"),
      ]);
    }, `Navigate a frame via ${navigationElement} with javascript:-urls in enforcing mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement,
        {defaultpolicy: 1, frame: 1});
      openWindow(t, `support/${joinToHref(params, kNavigationSupport)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupport),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, defaultpolicywashere: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a frame via ${navigationElement} with javascript:-urls w/ default policy in enforcing mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement, {frame: 1})
      openWindow(t, `support/${joinToHref(params, kNavigationSupportReportOnly)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupportReportOnly),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, continue: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a frame via ${navigationElement} with javascript:-urls in report-only mode.`);

    promise_test(t => {
      const params = maybeAddFormSubmissionToSearchParams(navigationElement,
                                                          {defaultpolicy: 1, frame: 1});
      openWindow(t, `support/${joinToHref(params, kNavigationSupportReportOnly)}`);
      return Promise.all([
        expectLoadedAsMessage(params, kNavigationSupportReportOnly),
        expectLoadedAsMessage(Object.assign({navigationattempted: 1, defaultpolicywashere: 1}, params), kNavigationSupport),
      ]);
    }, `Navigate a frame via ${navigationElement} with javascript:-urls w/ default policy in report-only mode.`);
  }
</script>
</body>
